Create api.py
Flask:  Flask is a lightweight web application framework in Python.
jsonify: This function from Flask is used to convert Python dictionaries to JSON responses
pandas as pd: Pandas is a powerful data manipulation library in Python, often used for data analysis and handling structured data. It's imported as pd for convenience.
app: This creates an instance of the Flask class. The __name__ argument helps Flask determine the root path of the application.