Create app.js
React: This is the main library for building user interfaces in a component-based architecture.
useEffect: This is a React Hook that allows you to perform side effects in function components, such as fetching data from an API or subscribing to a data stream.
useState: This is another React Hook that allows you to add state to function components.
function App() { ... }: This defines a functional component named App.
const [foodTrucks, setFoodTrucks] = useState([]): This line declares a state variable foodTrucks and a setter function setFoodTrucks. The initial state is set to an empty array. foodTrucks will store the list of food trucks, and setFoodTrucks will be used to update this state.
useEffect(() => { ... }, []);: This hook is used to perform side effects. In this case, it fetches data from the /api/foodtrucks endpoint. The empty array [] as the second argument means this effect will only run once after the initial render.
fetch('/api/foodtrucks'): This uses the Fetch API to make a GET request to the /api/foodtrucks endpoint.
.then(response => response.json()): This processes the response and converts it to JSON.
.then(data => setFoodTrucks(data)): This takes the data from the response and updates the foodTrucks state using the setFoodTrucks setter function
return ( ... ): This is the JSX returned by the App component. JSX is a syntax extension that looks similar to HTML but is used with React to describe the UI structure.
<div> ... </div>: This is a wrapper div for the component's content.
<h1>Food Trucks in San Francisco</h1>: This is a header for the list of food trucks.
<ul> ... </ul>: This is an unordered list that will contain the food trucks.
{foodTrucks.map(truck => ( ... ))}: This iterates over the foodTrucks array using the map function. For each truck, it returns a list item (<li>).
<li key={truck.id}>{truck.name}</li>: Each list item displays the name of the food truck. The key attribute is used to give each item a unique identifier, which helps React efficiently update the list.